+++ lib/src/controllers/mp_undo_redo_controller.dart
@@ -4,6 +4,7 @@ import 'package:mapiah/src/commands/types/mp_command_description_type.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/elements/th_file.dart';
 import 'package:mapiah/src/commands/mp_undo_redo_command.dart';
+import 'package:mapiah/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart';
 import 'package:mobx/mobx.dart';
 
 part 'mp_undo_redo_controller.g.dart';
@@ -122,6 +123,10 @@ abstract class MPUndoRedoControllerBase with Store {
 
     command.execute(_th2FileEditController);
 
+    _th2FileEditController.stateController.setState(
+      MPTH2FileEditStateType.selectEmptySelection,
+    );
+
     _redos.add(redo);
     _th2FileEditController.triggerAllElementsRedraw();
   }
@@ -139,6 +144,10 @@ abstract class MPUndoRedoControllerBase with Store {
     );
 
     command.execute(_th2FileEditController);
+
+    _th2FileEditController.stateController.setState(
+      MPTH2FileEditStateType.selectEmptySelection,
+    );
 
     _th2FileEditController.triggerAllElementsRedraw();
     _undos.add(undo);